refactor(backend): restructure backend application
This commit introduces a major refactoring of the backend application
to improve modularity and align with NestJS best practices. The new
structure organizes the application into distinct modules and a shared kernel.

Key changes:
- **Modularization**: Business logic is now organized into modules (`auth`, `health`).
- **Shared Kernel**: Common functionalities like decorators, exceptions,
  and the Prisma module are moved to a `shared` directory.
- **Prisma Schema**: The Prisma schema path in the documentation has been
  updated to reflect the new directory-based structure.
- **Auth Module**: The authentication logic has been encapsulated within
  its own module.
