+++ .commits/refactor-backend-split-prisma-schema-2026-01-06.txt
@@ -1,0 +1,10 @@
+refactor(backend): split prisma schema into multiple files
+
+The Prisma schema has been refactored from a single `schema.prisma` file
+into a directory of smaller, model-specific files.
+
+This modular approach improves organization and maintainability of the
+database schema, especially as the number of models grows.
+
+The `apps/backend/package.json` has been updated to point to the new schema
+directory.
