+++ apps/backend/src/app.module.ts
@@ -3,12 +3,12 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { configuration, validationSchema } from './config/configuration';
 import { ConfigModule } from '@nestjs/config';
-import { PrismaModule } from './prisma/prisma.module';
-import { HealthController } from './health/health.controller';
-import { AuthModule } from './auth/auth.module';
+import { PrismaModule } from './infra/prisma/prisma.module';
+import { HealthController } from './modules/health/health.controller';
+import { AuthModule } from './modules/auth/auth.module';
 import { APP_FILTER, APP_GUARD } from '@nestjs/core';
-import { SecurityExceptionFilter } from './exceptions/security-exception.filter';
-import { GlobalAuthGuard } from './auth/global-auth.guard';
+import { SecurityExceptionFilter } from './shared/exceptions/security-exception.filter';
+import { GlobalAuthGuard } from './modules/auth/global-auth.guard';
 
 @Module({
   imports: [
