+++ apps/backend/src/infra/prisma/prisma.module.ts
@@ -1,3 +1,5 @@
+// src/infra/prisma/prisma.module.ts
+
 import { Global, Module } from '@nestjs/common';
 import { PrismaService } from './prisma.service';
