+++ apps/backend/src/modules/auth/auth.module.ts
@@ -2,7 +2,7 @@
 
 import { Module } from '@nestjs/common';
 import { UserService } from './user/user.service';
-import { PrismaService } from '../prisma/prisma.service';
+import { PrismaService } from '../../infra/prisma/prisma.service';
 import { GlobalAuthGuard } from './global-auth.guard';
 
 @Module({
