+++ apps/backend/src/modules/auth/user/user.service.ts
@@ -2,7 +2,7 @@
 
 import { Injectable } from '@nestjs/common';
 import { User } from '@prisma/client';
-import { PrismaService } from 'src/prisma/prisma.service';
+import { PrismaService } from 'src/infra/prisma/prisma.service';
 
 @Injectable()
 export class UserService {
