+++ apps/backend/src/shared/exceptions/security-exception.filter.ts
